tests/wpt/tests/html/canvas/element/line-styles/2d.line.join.open.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.line.join.open</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.join.open</h1>
<p class="desc">Line joins are not drawn at the corner of an unclosed rectangle</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Line joins are not drawn at the corner of an unclosed rectangle");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#0f0';
  ctx.strokeStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.lineJoin = 'miter';
  ctx.lineWidth = 200;

  ctx.beginPath();
  ctx.moveTo(100, 50);
  ctx.lineTo(100, 1000);
  ctx.lineTo(1000, 1000);
  ctx.lineTo(1000, 50);
  ctx.lineTo(100, 50);
  ctx.stroke();

  _assertPixel(canvas, 1,1, 0,255,0,255);
  _assertPixel(canvas, 48,1, 0,255,0,255);
  _assertPixel(canvas, 48,48, 0,255,0,255);
  _assertPixel(canvas, 1,48, 0,255,0,255);

});
</script>
